Update patch set 1
Patch Set 1: No score; I would prefer that you didn't submit this

This change introduces quite a different behavior. When the activity is paused (for instance if you press home or switch to another window) it is not destroyed anymore, which will cause reactivation problems when the user goes back to it (it will be resumed but everything was closed in onPause().)

You need a better solution than this.

Patch-set: 1
Reviewer: Romain Guy <1001146@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1
Label: Verified=0